Edit contact: save contact to SIM!
The new/edited contact is now saved to SIM. We wait the respone, if OK then we go switch to last screen, if ERROR then we wtay on the contact edit screen.

Added TRIM_STRING and TRIM_STRING_SPACES to make triming strings to a certain length easier. Replaced in several places in the code.

A setting allows to choose where from to load the contacts (sim, code, sd)

Fixed various glitches in relation to contacts