Add Quote Mode to parser
Added a Quote Mode to parser. When we encounter a first " in the modem line, quote mode is turned on, when we encounter the second " quote mode is turned off. When in quote mode, all the characters that are usually used for parsing (: and ,) are added to the current field instead.