SMS: get SMS of current contact; Country Code and numbers comparison
We first go through all SMS's numbers. For SMS with the number of current contact, we add it to an array. Then, for every SMS in this array: if it is present in cache, we don't have to do anything; if it is not, we query it from modem and add it to cache.

Added the Contact SMS pane for display.

Also added a Country Code parameter, and a subroutine to compare two phone numbers (taking into account the country code).