SMS: used default reporting for new SMS
We use the default reporting of new messages: an URC (+CMTI) gives us the index of the newly stored message, which we query afterwards. This will allow messages received when software is not running to be queried later.

Added a dd flag, used to enable debugging at different stages (related to SMS), instead of having to manually set debugging in those stages.